* bugfix: making pm_id as uint32_t (#147)


* increased max kernel arg to 512 (#157)


* added check if eof is missing in asm (#156)


* fix cp patch schema in aie2ps (#159)


* bd len fix (#160)


* Add workflow to pull main-ge from github.com (#161)

The workflow overwrites this repository's main-ge branch with
https://github.com/Xilinx/aiebu/tree/main-ge


* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Change reset and force push to rebase and push


* Update pull_main-ge.yml

* Update pull_main-ge.yml

* Update pull_main-ge.yml

* remove pull_main-ge.yml


---------

+++ src/cpp/aiebu/src/preprocessor/preprocessor_input.h
@@ -27,7 +27,7 @@ public:
                         const std::vector<char>&,
                         const std::vector<std::string>&,
                         const std::vector<std::string>&,
-                        const std::map<uint8_t, std::vector<char> >& ctrlpkt) = 0;
+                        const std::map<uint32_t, std::vector<char> >& ctrlpkt) = 0;
 
   const std::vector<std::string> get_keys()
   {
